Check annex version before using annex.private in clone
This introduces a version kludge indicating whether private mode is
supported by git-annex and adjusts tests for clone to account for it.

Furthermore this patch makes `clone` fail if private mode was
requested by command parameter but isn't supported by installed
git-annex. This seems preferrable over a fallback solution like `git
annex dead here`, since the behavior is not the same and whether such
an approach suits the usecase is up to the user to decide.

Note. This is a slimmed-down version of the commit originally created
by Ben. Ben's version also dealt with create. We have since decided to
only cover clone, not create. Personally, I think we should have
bumped the minimal git version a while ago (removing the need for the
kludge), but this commit is neither the time nor the place to do it;
ping #6847 -- MSz.

Co-authored-by: Benjamin Poldrack <benjaminpoldrack@gmail.com>